refactor: shared utils
