+++ package.json
@@ -102,7 +102,6 @@
     "qrcode": "^1.5.1",
     "rxdb": "^13.15.3",
     "rxjs": "^7.8.0",
-    "scripts": "workspace:*",
     "shared": "workspace:*",
     "signature_pad": "^4.1.5",
     "stacktrace-js": "^2.0.2",
