+++ packages/data-models/package.json
@@ -10,7 +10,6 @@
   "devDependencies": {
     "@types/clone": "^2.1.1",
     "@types/node": "^16.18.9",
-    "scripts": "workspace:*",
     "shared": "workspace:*",
     "typescript": "~4.2.4"
   },
