+++ packages/scripts/package.json
@@ -42,6 +42,7 @@
     "ts-morph": "^15.0.0",
     "typescript": "^4.1.2",
     "winston": "^3.8.1",
+    "workflows": "workspace:*",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
@@ -51,6 +52,7 @@
     "@types/jasmine": "^3.10.6",
     "@types/node-rsa": "^1.1.1",
     "@types/semver": "^7.3.9",
+    "@types/winston": "^2.4.4",
     "esbuild": "^0.14.2",
     "jasmine": "^3.99.0",
     "jasmine-spec-reporter": "^7.0.0",
