+++ packages/scripts/src/commands/app-data/convert/cacheStrategy/jsonFile.spec.ts
@@ -2,7 +2,7 @@ import { emptyDirSync, existsSync, readdirSync, readJsonSync, rmdirSync, statSyn
 import path from "path";
 import { JsonFileCache } from "./jsonFile";
 
-import { SCRIPTS_WORKSPACE_PATH } from "../../../../paths";
+import { SCRIPTS_WORKSPACE_PATH } from "shared";
 const testCacheDir = path.resolve(SCRIPTS_WORKSPACE_PATH, "test", "data", "cache", "spec");
 
 const testData = {
