+++ packages/scripts/src/commands/app-data/convert/processors/flowParser/flowParser.spec.ts
@@ -1,6 +1,6 @@
 import path from "path";
 import { FlowTypes } from "data-models";
-import { SCRIPTS_WORKSPACE_PATH } from "../../../../../paths";
+import { SCRIPTS_WORKSPACE_PATH } from "shared";
 import { clearLogs, getLogs } from "../../utils";
 import { FlowParserProcessor } from "./flowParser";
