+++ packages/scripts/src/commands/app-data/convert/utils/index.ts
@@ -3,4 +3,4 @@ export * from "./app-data-condition.utils";
 export * from "./app-data-action.utils";
 export * from "./logging";
 // Re-export common shared utils for easier import
-export * from "../../../../utils";
+export * from "shared";
