+++ packages/scripts/src/commands/app-data/postProcess/sheets.spec.ts
@@ -6,7 +6,7 @@ import mockFs from "mock-fs";
 
 // Use default imports to allow spying on functions and replacing with mock methods
 import { ActiveDeployment } from "../../deployment/get";
-import { Logger } from "../../../utils/logging.utils";
+import { Logger } from "shared";
 import path from "path";
 
 /** Mock file system folders for use in tests */
