+++ packages/scripts/src/commands/deployment/create.ts
@@ -2,8 +2,8 @@
 import { Command } from "commander";
 import fs from "fs-extra";
 import path from "path";
-import { DEPLOYMENTS_PATH } from "../../paths";
-import { Logger, logOutput, logWarning, promptInput, promptOptions } from "../../utils";
+import { DEPLOYMENTS_PATH } from "shared";
+import { Logger, logOutput, logWarning, promptInput, promptOptions } from "shared";
 import type { IDeploymentConfigJson } from "./common";
 import { DeploymentSet } from "./set";
 import generateDefaultConfig from "./templates/config.default";
