+++ packages/scripts/src/commands/deployment/get.ts
@@ -3,8 +3,8 @@ import { Command } from "commander";
 import { IDeploymentConfig } from "data-models";
 import fs from "fs-extra";
 import path from "path";
-import { DEPLOYMENTS_PATH } from "../../paths";
-import { Logger } from "../../utils";
+import { DEPLOYMENTS_PATH } from "shared";
+import { Logger } from "shared";
 import type { IDeploymentConfigJson } from "./common";
 import { loadTSFileDefaultExport } from "./compile";
 import { loadDeploymentJson } from "./utils";
