+++ packages/scripts/src/commands/deployment/set.ts
@@ -2,8 +2,8 @@
 import { Command } from "commander";
 import fs from "fs-extra";
 import path from "path";
-import { DEPLOYMENTS_PATH, ROOT_DIR } from "../../paths";
-import { promptOptions, logOutput, logWarning } from "../../utils";
+import { DEPLOYMENTS_PATH, ROOT_DIR } from "shared";
+import { promptOptions, logOutput, logWarning } from "shared";
 import { ActiveDeployment } from "./get";
 import { loadDeploymentJson } from "./utils";
