+++ packages/scripts/src/commands/e2e-data/parse.ts
@@ -5,7 +5,7 @@ import XLSX from "xlsx";
 import prettier from "prettier";
 import * as mappings from "./mappings";
 import snippets from "./snippets";
-import { ROOT_DIR } from "../../paths";
+import { ROOT_DIR } from "shared";
 
 interface IProgramOptions {
   input: string;
