+++ packages/scripts/src/tasks/providers/android.ts
@@ -1,8 +1,7 @@
 import * as path from "path";
 import { Options, run } from "cordova-res";
 import fs from "fs";
-import { ROOT_DIR } from "../../paths";
-import { Logger } from "../../utils";
+import { ROOT_DIR, Logger } from "shared";
 
 const set_splash_image = async (splashAssetPath: string) => {
   if (!fs.existsSync(splashAssetPath)) {
