+++ packages/scripts/src/tasks/providers/encryption.ts
@@ -11,8 +11,8 @@ import {
 import NodeRSA from "node-rsa";
 import { basename, resolve } from "path";
 import { ActiveDeployment } from "../../commands/deployment/get";
-import { DEPLOYMENTS_PATH } from "../../paths";
-import { logOutput, promptConfirmation, promptEditorInput } from "../../utils";
+import { DEPLOYMENTS_PATH } from "shared";
+import { logOutput, promptConfirmation, promptEditorInput } from "shared";
 
 /** Suffix added to all encrypted files to distinguish from originals */
 const ENCRYPTED_SUFFIX = "crypt";
