+++ packages/scripts/src/tasks/providers/subtitles.ts
@@ -2,8 +2,8 @@ import chalk from "chalk";
 import fs from "fs";
 import * as path from "path";
 import parser from "subtitles-parser-vtt";
-import { ROOT_DIR } from "../../paths";
-import { recursiveFindByExtension, Logger, logOutput, promptOptions } from "../../utils";
+import { ROOT_DIR } from "shared";
+import { recursiveFindByExtension, Logger, logOutput, promptOptions } from "shared";
 
 // This is an interim solution for manually generating translated .vtt files,
 // to be used pending reworking the translations repo/pipeline to handle assets more generally
