--- packages/scripts/src/utils/index.ts
@@ -1,5 +1,0 @@
-export * from "./cli-utils";
-export * from "./file-utils";
-export * from "./logging.utils";
-export * from "./string-utils";
-export * from "./typescript-utils";
