+++ packages/scripts/tsconfig.json
@@ -8,7 +8,7 @@
     // Global type definitions.
     "typeRoots": ["./node_modules/@types"]
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "../shared/src/utils/file-utils.ts"],
   "ts-node": {
     "swc": true
   }
