+++ packages/shared/README.md
@@ -1,0 +1,6 @@
+## Shared
+
+Common platform data structures and manipulation methods.
+
+Note, some methods will only work in node-based environments (non-browser)
+
