+++ packages/shared/src/index.ts
@@ -1,2 +1,8 @@
-export * from "./utils";
 export * from "./models";
+export * from "./types";
+export * from "./utils";
+export * from "./paths";
+
+// Re-export all paths within a single namespace
+import * as PATHS from "./paths";
+export { PATHS };
