+++ packages/shared/src/paths.ts
