+++ packages/shared/src/utils/misc-utils.ts
@@ -1,9 +1,11 @@
 import { spawnSync } from "child_process";
-import { ROOT_DIR } from "shared";
+import { ROOT_DIR } from "../paths";
 
-const runTSCodeFormatter = (folderPath: string) => {
+/**
+ * Run prettier to automatically format code in given folder path
+ * NOTE - by default will only format .ts files
+ */
+export function runPrettierCodeTidy(folderPath: string) {
   const cmd = `npx prettier --config ${ROOT_DIR}/.prettierrc --write ${folderPath}/**/*.ts --loglevel error`;
   return spawnSync(cmd, { stdio: ["inherit", "inherit", "inherit"], shell: true });
-};
-
-export default { runTSCodeFormatter };
+}
