+++ packages/shared/src/utils/string-utils.ts
