+++ packages/shared/src/utils/typescript-utils.ts
