+++ packages/workflows/README.md
@@ -1,0 +1,7 @@
+## Workflows
+
+Common workflows made available to scripts
+
+```sh
+yarn workflow [name_of_workflow]
+```
