+++ packages/workflows/package.json
@@ -7,7 +7,6 @@
     "build": "tsc -b",
     "serve": "tsc -w"
   },
-  "dependencies": {},
   "devDependencies": {
     "@types/node": "^16.18.9",
     "shared": "workspace:*",
